Flip the arguments of move
This is a good idea in general --- you always want to try to line up
function arguments like this. If a function returns the same type as
one of its arguments, that argument should be last. Doing so will make
your life significantly easier when working with partially applied
functions.